test_requirements.txt
pytest
pytest-cov
python-coveralls
pytest-flask
pytest-xdist
codacy-coverage
